requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0

# Vector store and embeddings
qdrant-client>=1.1.1
sentence-transformers>=2.2.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.9.3
markdown>=3.3.4

# LLM and processing
openai>=0.27.0
langchain>=0.0.200
numpy>=1.21.0
pandas>=1.3.0

# Testing and development
pytest>=7.0.0
black>=22.3.0
flake8>=4.0.0

# API and async
httpx>=0.23.0
aiohttp>=3.8.0
pydantic>=1.9.0

# Together.ai package
together>=0.1.8